+++ CHANGELOG.md
@@ -11,6 +11,8 @@ For more details or to discuss releases, please visit the
 
 ## Next
 
+- don't require poll period to be set for modbus server
+
 ## [[0.0.27] - 2021-04-15](https://github.com/simpleiot/simpleiot/releases/tag/v0.0.27)
 
 - slow down manual scanning to reduce background CPU usage